test(voice): re-add UIState import after rebase
