+++ docs/reference/configuration.md
@@ -483,25 +483,27 @@ their corresponding top-level category object in your `settings.json` file.
   - **Description:** Skip the next speaker check.
   - **Default:** `true`
 
-#### `voice`
+#### `experimental.voice`
 
-- **`voice.activationMode`** (enum):
+Settings for voice mode and transcription.
+
+- **`experimental.voice.activationMode`** (enum):
   - **Description:** How to trigger voice recording with the Space key.
   - **Default:** `"push-to-talk"`
   - **Values:** `"push-to-talk"`, `"toggle"`
 
-- **`voice.backend`** (enum):
+- **`experimental.voice.backend`** (enum):
   - **Description:** The backend to use for voice transcription.
   - **Default:** `"gemini-live"`
   - **Values:** `"gemini-live"`, `"whisper"`
 
-- **`voice.whisperModel`** (enum):
+- **`experimental.voice.whisperModel`** (enum):
   - **Description:** The Whisper model to use for local transcription.
   - **Default:** `"ggml-base.en.bin"`
   - **Values:** `"ggml-tiny.en.bin"`, `"ggml-base.en.bin"`,
     `"ggml-large-v3-turbo-q5_0.bin"`, `"ggml-large-v3-turbo-q8_0.bin"`
 
-- **`voice.stopGracePeriodMs`** (number):
+- **`experimental.voice.stopGracePeriodMs`** (number):
   - **Description:** How long to wait for final transcription after stopping
     recording.
   - **Default:** `1000`
