+++ packages/cli/src/config/settingsSchema.ts
@@ -1115,75 +1115,6 @@ const SETTINGS_SCHEMA = {
     },
   },
 
-  voice: {
-    type: 'object',
-    label: 'Voice',
-    category: 'Voice',
-    requiresRestart: false,
-    default: {},
-    description: 'Settings for voice mode and transcription.',
-    showInDialog: true,
-    properties: {
-      activationMode: {
-        type: 'enum',
-        label: 'Activation Mode',
-        category: 'Voice',
-        requiresRestart: false,
-        default: 'push-to-talk',
-        description: 'How to trigger voice recording with the Space key.',
-        showInDialog: true,
-        options: [
-          { value: 'push-to-talk', label: 'Push-To-Talk (Hold Space)' },
-          { value: 'toggle', label: 'Toggle (Press Space to start/stop)' },
-        ],
-      },
-      backend: {
-        type: 'enum',
-        label: 'Transcription Backend',
-        category: 'Voice',
-        requiresRestart: false,
-        default: 'gemini-live',
-        description: 'The backend to use for voice transcription.',
-        showInDialog: true,
-        options: [
-          { value: 'gemini-live', label: 'Gemini Live API (Cloud)' },
-          { value: 'whisper', label: 'Whisper (Local)' },
-        ],
-      },
-      whisperModel: {
-        type: 'enum',
-        label: 'Whisper Model',
-        category: 'Voice',
-        requiresRestart: false,
-        default: 'ggml-base.en.bin',
-        description: 'The Whisper model to use for local transcription.',
-        showInDialog: true,
-        options: [
-          { value: 'ggml-tiny.en.bin', label: 'Tiny (EN) - Fast (~75MB)' },
-          { value: 'ggml-base.en.bin', label: 'Base (EN) - Balanced (~142MB)' },
-          {
-            value: 'ggml-large-v3-turbo-q5_0.bin',
-            label: 'Large v3 Turbo (Q5_0) - High Accuracy (~547MB)',
-          },
-          {
-            value: 'ggml-large-v3-turbo-q8_0.bin',
-            label: 'Large v3 Turbo (Q8_0) - Max Accuracy (~834MB)',
-          },
-        ],
-      },
-      stopGracePeriodMs: {
-        type: 'number',
-        label: 'Stop Grace Period (ms)',
-        category: 'Voice',
-        requiresRestart: false,
-        default: 1000,
-        description:
-          'How long to wait for final transcription after stopping recording.',
-        showInDialog: true,
-      },
-    },
-  },
-
   modelConfigs: {
     type: 'object',
     label: 'Model Configs',
@@ -2139,6 +2070,77 @@ const SETTINGS_SCHEMA = {
         description:
           'Enable experimental voice dictation and commands (/voice, /voice model).',
         showInDialog: true,
+      },
+      voice: {
+        type: 'object',
+        label: 'Voice',
+        category: 'Experimental',
+        requiresRestart: false,
+        default: {},
+        description: 'Settings for voice mode and transcription.',
+        showInDialog: false,
+        properties: {
+          activationMode: {
+            type: 'enum',
+            label: 'Voice Activation Mode',
+            category: 'Experimental',
+            requiresRestart: false,
+            default: 'push-to-talk',
+            description: 'How to trigger voice recording with the Space key.',
+            showInDialog: true,
+            options: [
+              { value: 'push-to-talk', label: 'Push-To-Talk (Hold Space)' },
+              { value: 'toggle', label: 'Toggle (Press Space to start/stop)' },
+            ],
+          },
+          backend: {
+            type: 'enum',
+            label: 'Voice Transcription Backend',
+            category: 'Experimental',
+            requiresRestart: false,
+            default: 'gemini-live',
+            description: 'The backend to use for voice transcription.',
+            showInDialog: true,
+            options: [
+              { value: 'gemini-live', label: 'Gemini Live API (Cloud)' },
+              { value: 'whisper', label: 'Whisper (Local)' },
+            ],
+          },
+          whisperModel: {
+            type: 'enum',
+            label: 'Whisper Model',
+            category: 'Experimental',
+            requiresRestart: false,
+            default: 'ggml-base.en.bin',
+            description: 'The Whisper model to use for local transcription.',
+            showInDialog: true,
+            options: [
+              { value: 'ggml-tiny.en.bin', label: 'Tiny (EN) - Fast (~75MB)' },
+              {
+                value: 'ggml-base.en.bin',
+                label: 'Base (EN) - Balanced (~142MB)',
+              },
+              {
+                value: 'ggml-large-v3-turbo-q5_0.bin',
+                label: 'Large v3 Turbo (Q5_0) - High Accuracy (~547MB)',
+              },
+              {
+                value: 'ggml-large-v3-turbo-q8_0.bin',
+                label: 'Large v3 Turbo (Q8_0) - Max Accuracy (~834MB)',
+              },
+            ],
+          },
+          stopGracePeriodMs: {
+            type: 'number',
+            label: 'Voice Stop Grace Period (ms)',
+            category: 'Experimental',
+            requiresRestart: false,
+            default: 1000,
+            description:
+              'How long to wait for final transcription after stopping recording.',
+            showInDialog: true,
+          },
+        },
       },
       adk: {
         type: 'object',
