+++ packages/cli/src/ui/components/InputPrompt.tsx
@@ -55,10 +55,8 @@ import {
   coreEvents,
   debugLogger,
   type Config,
-  AudioRecorder,
-  TranscriptionFactory,
-  type TranscriptionProvider,
 } from '@google/gemini-cli-core';
+import { useVoiceMode } from '../hooks/useVoiceMode.js';
 import {
   parseInputForHighlighting,
   parseSegmentsFromTokens,
@@ -162,9 +160,6 @@ export function isLargePaste(text: string): boolean {
 }
 
 const DOUBLE_TAB_CLEAN_UI_TOGGLE_WINDOW_MS = 350;
-const HOLD_DELAY_MS = 600;
-const RELEASE_DELAY_MS = 300;
-
 /**
  * Attempt to toggle expansion of a paste placeholder in the buffer.
  * Returns true if a toggle action was performed or hint was shown, false otherwise.
@@ -270,7 +265,7 @@ export const InputPrompt: React.FC<InputPromptProps> = ({
           resetEscapeState();
           if (buffer.text.length > 0) {
             buffer.setText('');
-            turnBaselineRef.current = null;
+            resetTurnBaseline();
             resetCompletionState();
           } else if (history.length > 0) {
             onSubmit('/rewind');
@@ -289,238 +284,15 @@ export const InputPrompt: React.FC<InputPromptProps> = ({
   const hasUserNavigatedSuggestions = useRef(false);
   const listRef = useRef<ScrollableListRef<ScrollableItem>>(null);
 
-  const [isRecording, setIsRecording] = useState(false);
-  const liveTranscriptionRef = useRef('');
-  const stopRequestedRef = useRef(false);
-  const isConnectingRef = useRef(false);
-  const isRecordingRef = useRef(false); // Track logic state independently of UI state
-  const lastFailureTimeRef = useRef(0);
-  const recordingInProgressRef = useRef(false); // Lock for toggle
-  const voiceTimeoutRef = useRef<NodeJS.Timeout | null>(null);
-  const recorderRef = useRef<AudioRecorder | null>(null);
-  const transcriptionServiceRef = useRef<TranscriptionProvider | null>(null);
-  const turnBaselineRef = useRef<string | null>(null);
-
-  const pttStateRef = useRef<'idle' | 'possible-hold' | 'recording'>('idle');
-  const pttTimerRef = useRef<NodeJS.Timeout | null>(null);
-  const disconnectTimerRef = useRef<NodeJS.Timeout | null>(null);
-
-  const bufferRef = useRef(buffer);
-  bufferRef.current = buffer;
-
-  const stopVoiceRecording = useCallback(() => {
-    debugLogger.debug('[Voice] Stop requested');
-    stopRequestedRef.current = true;
-    setIsRecording(false);
-    isRecordingRef.current = false;
-    isConnectingRef.current = false;
-
-    if (recorderRef.current) {
-      recorderRef.current.stop();
-      recorderRef.current = null;
-    }
-
-    const serviceToDisconnect = transcriptionServiceRef.current;
-    transcriptionServiceRef.current = null; // Detach immediately so new session can start if needed
-
-    if (serviceToDisconnect) {
-      const gracePeriodMs = settings.voice?.stopGracePeriodMs ?? 1000;
-      debugLogger.debug(
-        `[Voice] Draining transcription for ${gracePeriodMs}ms`,
-      );
-
-      if (disconnectTimerRef.current) clearTimeout(disconnectTimerRef.current);
-      disconnectTimerRef.current = setTimeout(() => {
-        debugLogger.debug('[Voice] Grace period ended, disconnecting service');
-        serviceToDisconnect.disconnect();
-        disconnectTimerRef.current = null;
-      }, gracePeriodMs);
-    }
-
-    liveTranscriptionRef.current = '';
-    pttStateRef.current = 'idle';
-  }, [settings.voice]);
-
-  const startVoiceRecording = useCallback(() => {
-    // Check for cooldown after failure
-    if (Date.now() - lastFailureTimeRef.current < 2000) {
-      return;
-    }
-
-    // Cancel any pending disconnect if we start a new session quickly
-    if (disconnectTimerRef.current) {
-      clearTimeout(disconnectTimerRef.current);
-      disconnectTimerRef.current = null;
-    }
-
-    recordingInProgressRef.current = true;
-    turnBaselineRef.current = bufferRef.current.text;
-
-    isConnectingRef.current = true;
-    setIsRecording(true);
-    isRecordingRef.current = true;
-
-    liveTranscriptionRef.current = '';
-    stopRequestedRef.current = false;
-
-    const apiKey =
-      config.getContentGeneratorConfig()?.apiKey ||
-      process.env['GEMINI_API_KEY'] ||
-      '';
-
-    const startAsync = async () => {
-      const cleanupIfStopped = () => {
-        if (stopRequestedRef.current) {
-          if (recorderRef.current) {
-            recorderRef.current.stop();
-            recorderRef.current = null;
-          }
-          if (transcriptionServiceRef.current) {
-            transcriptionServiceRef.current.disconnect();
-            transcriptionServiceRef.current = null;
-          }
-          setIsRecording(false);
-          isRecordingRef.current = false;
-          isConnectingRef.current = false;
-          recordingInProgressRef.current = false;
-
-          return true;
-        }
-        return false;
-      };
-
-      if (cleanupIfStopped()) return;
-
-      const voiceBackend = settings.voice?.backend ?? 'gemini-live';
-
-      if (!apiKey && voiceBackend === 'gemini-live') {
-        setQueueErrorMessage(
-          'Cloud voice mode requires a GEMINI_API_KEY. Please set it in your environment or ~/.gemini/.env.',
-        );
-        setIsRecording(false);
-        isRecordingRef.current = false;
-        isConnectingRef.current = false;
-        recordingInProgressRef.current = false;
-        lastFailureTimeRef.current = Date.now();
-        return;
-      }
-
-      if (voiceBackend === 'gemini-live') {
-        recorderRef.current = new AudioRecorder();
-      }
-
-      transcriptionServiceRef.current = TranscriptionFactory.createProvider(
-        settings.voice,
-        apiKey,
-      );
-
-      // TranscriptionProviders may be session-cumulative (full text so far) or
-      // turn-based (resetting text after turnComplete).
-      transcriptionServiceRef.current.on('transcription', (text) => {
-        if (text) {
-          const currentBufferText = bufferRef.current.text;
-          const previousTranscription = liveTranscriptionRef.current;
-
-          let newTotalText = currentBufferText;
-
-          // If we have a previous transcription chunk, find it at the end of the buffer and remove it
-          if (
-            previousTranscription &&
-            currentBufferText.endsWith(previousTranscription)
-          ) {
-            newTotalText = currentBufferText.slice(
-              0,
-              -previousTranscription.length,
-            );
-          } else if (
-            currentBufferText &&
-            !currentBufferText.endsWith(' ') &&
-            !currentBufferText.endsWith('\n')
-          ) {
-            // If this is the start of a new transcription and the buffer doesn't end with whitespace, add a space
-            newTotalText += ' ';
-          }
-
-          newTotalText += text;
-          bufferRef.current.setText(newTotalText, 'end');
-        }
-
-        liveTranscriptionRef.current = text;
-      });
-
-      transcriptionServiceRef.current.on('turnComplete', () => {
-        // When a turn is complete, some providers (like Gemini Live) will start a new
-        // transcription from empty.
-        liveTranscriptionRef.current = '';
-      });
-
-      transcriptionServiceRef.current.on('error', (err) => {
-        debugLogger.error('[Voice] Transcription error:', err);
-
-        lastFailureTimeRef.current = Date.now();
-        recordingInProgressRef.current = false;
-      });
-
-      transcriptionServiceRef.current.on('close', () => {
-        if (!stopRequestedRef.current) {
-          setIsRecording(false);
-          isRecordingRef.current = false;
-          isConnectingRef.current = false;
-          recordingInProgressRef.current = false;
-
-          lastFailureTimeRef.current = Date.now();
-        }
-      });
-
-      try {
-        await transcriptionServiceRef.current.connect();
-        if (cleanupIfStopped()) return;
-
-        await recorderRef.current?.start();
-        if (cleanupIfStopped()) return;
-
-        isConnectingRef.current = false; // Successfully connected
-
-        const voiceBackend = settings.voice?.backend ?? 'gemini-live';
-
-        recorderRef.current?.on('data', (chunk) => {
-          if (voiceBackend === 'gemini-live') {
-            transcriptionServiceRef.current?.sendAudioChunk(chunk);
-          }
-        });
-        recorderRef.current?.on('error', (err) => {
-          debugLogger.error('[Voice] Recorder error:', err);
-          setIsRecording(false);
-          isRecordingRef.current = false;
-          isConnectingRef.current = false;
-          recordingInProgressRef.current = false;
-
-          lastFailureTimeRef.current = Date.now();
-        });
-      } catch (err: unknown) {
-        const message = err instanceof Error ? err.message : String(err);
-        setQueueErrorMessage(`Voice mode failure: ${message}`);
-        setIsRecording(false);
-        isRecordingRef.current = false;
-        isConnectingRef.current = false;
-        recordingInProgressRef.current = false;
-
-        lastFailureTimeRef.current = Date.now();
-
-        // Cleanup refs on failure
-        if (recorderRef.current) {
-          recorderRef.current.stop();
-          recorderRef.current = null;
-        }
-        if (transcriptionServiceRef.current) {
-          transcriptionServiceRef.current.disconnect();
-          transcriptionServiceRef.current = null;
-        }
-      }
-    };
-
-    void startAsync();
-  }, [config, settings.voice, setQueueErrorMessage]);
+  const { isRecording, handleVoiceInput, resetTurnBaseline } = useVoiceMode({
+    buffer,
+    config,
+    settings,
+    setQueueErrorMessage,
+    isVoiceModeEnabled,
+    setVoiceModeEnabled,
+    keyMatchers,
+  });
 
   const [reverseSearchActive, setReverseSearchActive] = useState(false);
   const [commandSearchActive, setCommandSearchActive] = useState(false);
@@ -602,22 +374,11 @@ export const InputPrompt: React.FC<InputPromptProps> = ({
     }
   }, [showEscapePrompt, onEscapePromptChange]);
 
-  // Clear paste timeout and voice mode on unmount
+  // Clear paste timeout on unmount
   useEffect(
     () => () => {
       if (pasteTimeoutRef.current) {
         clearTimeout(pasteTimeoutRef.current);
-      }
-      if (voiceTimeoutRef.current) {
-        clearTimeout(voiceTimeoutRef.current);
-      }
-      if (recorderRef.current) {
-        recorderRef.current.stop();
-        recorderRef.current = null;
-      }
-      if (transcriptionServiceRef.current) {
-        transcriptionServiceRef.current.disconnect();
-        transcriptionServiceRef.current = null;
       }
     },
     [],
@@ -639,7 +400,7 @@ export const InputPrompt: React.FC<InputPromptProps> = ({
       // Clear the buffer *before* calling onSubmit to prevent potential re-submission
       // if onSubmit triggers a re-render while the buffer still holds the old value.
       buffer.setText('');
-      turnBaselineRef.current = null;
+      resetTurnBaseline();
       onSubmit(processedValue);
       resetCompletionState();
       resetReverseSearchCompletionState();
@@ -651,6 +412,7 @@ export const InputPrompt: React.FC<InputPromptProps> = ({
       shellModeActive,
       shellHistory,
       resetReverseSearchCompletionState,
+      resetTurnBaseline,
     ],
   );
 
@@ -900,101 +662,7 @@ export const InputPrompt: React.FC<InputPromptProps> = ({
 
   const handleInput = useCallback(
     (key: Key) => {
-      // While recording, suppress all keys except the space release (which is handled by useKeypress's nature)
-      // or Escape to cancel.
-      if (isRecording) {
-        const activationMode = settings.voice?.activationMode ?? 'push-to-talk';
-
-        if (keyMatchers[Command.ESCAPE](key)) {
-          stopVoiceRecording();
-          return true;
-        }
-
-        if (keyMatchers[Command.VOICE_MODE_PTT](key)) {
-          if (activationMode === 'push-to-talk') {
-            // In PTT mode, Space acts as a heartbeat. Reset the release timer.
-            if (pttTimerRef.current) {
-              clearTimeout(pttTimerRef.current);
-            }
-            pttTimerRef.current = setTimeout(() => {
-              stopVoiceRecording();
-              pttTimerRef.current = null;
-            }, RELEASE_DELAY_MS);
-            return true;
-          } else {
-            // In Toggle mode, Space stops the recording.
-            stopVoiceRecording();
-            return true;
-          }
-        }
-        // Swallow other keys during active recording to prevent keyboard interference
-        return true;
-      }
-
-      if (isVoiceModeEnabled) {
-        const activationMode = settings.voice?.activationMode ?? 'push-to-talk';
-
-        if (keyMatchers[Command.ESCAPE](key) && buffer.text === '') {
-          setVoiceModeEnabled(false);
-          return true;
-        }
-
-        if (keyMatchers[Command.VOICE_MODE_PTT](key)) {
-          if (
-            key.name === 'space' &&
-            !key.ctrl &&
-            !key.alt &&
-            !key.shift &&
-            !key.cmd
-          ) {
-            if (activationMode === 'toggle') {
-              startVoiceRecording();
-              return true;
-            } else {
-              // Push-To-Talk "Optimistic Space" Strategy
-              if (pttStateRef.current === 'idle') {
-                // First press: Insert space optimistically
-                buffer.insert(' ');
-                pttStateRef.current = 'possible-hold';
-
-                if (pttTimerRef.current) clearTimeout(pttTimerRef.current);
-                pttTimerRef.current = setTimeout(() => {
-                  // If no second space arrives within this window, it was a tap.
-                  pttStateRef.current = 'idle';
-                  pttTimerRef.current = null;
-                }, HOLD_DELAY_MS);
-                return true;
-              } else if (pttStateRef.current === 'possible-hold') {
-                // Second space arrives quickly (Key Repeat): It's a hold!
-                if (pttTimerRef.current) clearTimeout(pttTimerRef.current);
-
-                // Remove the optimistic space
-                buffer.backspace();
-
-                // Start recording
-                pttStateRef.current = 'recording';
-                startVoiceRecording();
-
-                // Start the release detection timer
-                pttTimerRef.current = setTimeout(() => {
-                  stopVoiceRecording();
-                  pttTimerRef.current = null;
-                }, RELEASE_DELAY_MS);
-                return true;
-              }
-            }
-          }
-        }
-
-        // If any other key is pressed during possible-hold, it was a fast type.
-        if (pttStateRef.current === 'possible-hold') {
-          pttStateRef.current = 'idle';
-          if (pttTimerRef.current) {
-            clearTimeout(pttTimerRef.current);
-            pttTimerRef.current = null;
-          }
-        }
-      }
+      if (handleVoiceInput(key)) return true;
 
       // Determine if this keypress is a history navigation command
       const isHistoryUp =
@@ -1222,10 +890,9 @@ export const InputPrompt: React.FC<InputPromptProps> = ({
       ) {
         setShellModeActive(!shellModeActive);
         buffer.setText(''); // Clear the '!' from input
-        turnBaselineRef.current = null;
+        resetTurnBaseline();
         return true;
       }
-
       if (keyMatchers[Command.ESCAPE](key)) {
         const cancelSearch = (
           setActive: (active: boolean) => void,
@@ -1710,6 +1377,7 @@ export const InputPrompt: React.FC<InputPromptProps> = ({
       backgroundTaskHeight,
       streamingState,
       handleEscPress,
+      resetTurnBaseline,
       registerPlainTabPress,
       resetPlainTabPress,
       toggleCleanUiDetailsVisible,
@@ -1719,14 +1387,9 @@ export const InputPrompt: React.FC<InputPromptProps> = ({
       keyMatchers,
       isHelpDismissKey,
       settings,
-      isVoiceModeEnabled,
-      isRecording,
-      setVoiceModeEnabled,
-      startVoiceRecording,
-      stopVoiceRecording,
+      handleVoiceInput,
     ],
   );
-
   useKeypress(handleInput, {
     isActive: !isEmbeddedShellFocused && !copyModeEnabled,
     priority: true,
@@ -2156,8 +1819,8 @@ export const InputPrompt: React.FC<InputPromptProps> = ({
               <Box flexDirection="row" marginBottom={0}>
                 <Text color={theme.text.secondary}>
                   &gt; Voice mode:{' '}
-                  {(settings.voice?.activationMode ?? 'push-to-talk') ===
-                  'push-to-talk'
+                  {(settings.experimental.voice?.activationMode ??
+                    'push-to-talk') === 'push-to-talk'
                     ? 'Hold Space to record'
                     : 'Space to start/stop recording'}{' '}
                   (Esc to exit)
